+++ frontend/lib/components/account_menu.dart
@@ -1,11 +1,11 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:nocsis/components/personal/user_avatar.dart';
+import 'package:nocsis/pages/classroom.dart';
 import 'package:nocsis/pages/console/index.dart';
 import 'package:nocsis/pages/licenses.dart';
 import 'package:nocsis/pages/settings/index.dart';
 import 'package:nocsis/routes/router.dart';
-import 'package:nocsis/screens/classroom.dart';
 
 class AccountMenu extends StatelessWidget {
   const AccountMenu({super.key});
